Factor solaris out
