+++ network/tests/fixtures/solaris/netstat_s_ptcp
@@ -1,0 +1,8 @@
+TCP: tcpRtoAlgorithm=     4 tcpRtoMin           =   200
+tcpRtoMax           = 60000 tcpMaxConn          =    -1
+tcpActiveOpens      =    57 tcpPassiveOpens     =    50
+tcpAttemptFails     =     1 tcpEstabResets      =     0
+tcpCurrEstab        =     0 tcpOutSegs          =   254
+tcpOutDataSegs      =   995 tcpOutDataBytes     =1216733
+tcpRetransSegs      =     0 tcpRetransBytes     =     0
+tcpOutAck           =   185 tcpOutAckDelayed    =     4
